initial tests done
