+++ docker-compose.test.yml
@@ -28,6 +28,20 @@ services:
       - BACKEND_PROXY=http://test-backend:8080
     networks:
       - ttadp-test
+  test-worker:
+    build:
+      context: .
+    container_name: test-worker
+    command: yarn server
+    env_file: .env
+    depends_on:
+      - test-db
+    environment:
+      - POSTGRES_HOST=test-db
+    volumes:
+      - "./worker:/app:rw"
+    networks:
+      - ttadp-test
   test-db:
     image: postgres:12.4
     container_name: test-db
