+++ worker/package.json
@@ -32,6 +32,9 @@
     "winston": "^3.3.3"
   },
   "devDependencies": {
+    "cross-env": "^7.0.3",
+    "jest": "^26.6.3",
+    "jest-junit": "^12.0.0",
     "nodemon": "^2.0.7"
   }
 }
