--- worker/src/worker.test.js
@@ -1,19 +1,0 @@
-import processFile from './worker';
-
-describe('File Upload', () => {
-  beforeAll(async () => {
-  });
-  afterAll(async () => {
-  });
-  afterEach(() => {
-    jest.clearAllMocks();
-  });
-
-  describe('File Upload Handlers happy path', () => {
-    beforeEach(() => {
-    });
-    it('tests a file upload', async () => {
-
-    });
-  });
-});
